+++ CHANGELOG.md
@@ -7,9 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0]
+
 ### Added
 
-* Now compatible with `pdm>=v2.8.2`.
+* Now compatible with `pdm>=v2.9.1`.
+* `pdm init` now shows Conda managed interpreters if runner specifier.
+* `pdm venv` commands work with Conda managed environments.
 * Now PyPI-Conda mapping url can be set with `conda.pypi-mapping.url` config.
 
 ### Changed